feat(epic-13): Refactor REST handlers para `_handler.ts` extraídos [wave 2]
Pré-requisito do MCP server (S-13.03/04). Cada Route Handler em
`app/api/v1/<resource>/route.ts` virou thin wrapper (auth + Zod + ok/fail);
a lógica de DB + audit + emit_event vive agora em `_handler.ts` reusável.

Tipo Actor discriminado em `lib/api/handlers/types.ts` permite que o mesmo
handler atenda usuário humano (cookie session) ou agente de IA (Bearer token,
S-13.03). Audit log usa actor_user_id quando type='user' e registra ator IA
em metadata.actor_type/actor_id quando type='ai_agent' — coluna dedicada
fica para wave futura, evitando migration nesta.

Handlers extraídos:
  - contacts: list, get, create, patch
  - conversations: list, get, updateStatus
  - messages: list, send (sent_via='ai' quando ator é ai_agent)
  - leads: list, get, create, update
  - pipelines: list (apenas _handler para MCP; sem Route REST no MVP)

Sub-rotas (leads bulk/move/win/lose, conversations claim/close/...) NÃO foram
refatoradas — só as 5 famílias com contracts emitidos pelo epic.

Comportamento REST preservado bit-a-bit: status codes, error codes, response
shapes, audit actions e event types idênticos. Suite REST E2E roda no QA.

Typecheck zero. Lint sem regressões.
