+++ app/api/v1/contacts/[id]/route.ts
@@ -1,34 +1,20 @@
 /**
- * GET   /api/v1/contacts/[id] — fetch single contact (RLS-scoped).
- *   Optional CPF decrypt: requires header `X-Decrypt-Purpose: <purpose>` AND
- *   role >= manager (Spec 05 / LGPD §4). Without header the response just
- *   advertises `cpf_available`.
+ * GET   /api/v1/contacts/[id] — fetch single (handler em ../_handler.ts)
+ * PATCH /api/v1/contacts/[id] — update (handler em ../_handler.ts)
  *
- * PATCH /api/v1/contacts/[id] — update mutable fields. Blocked when
- *   `is_anonymized=true` (LGPD irreversibility — code `lgpd_anonymization_irreversible`).
+ * Thin wrapper: auth + Zod + ok/fail. Decrypt CPF + LGPD irreversibility no handler.
  */
 import { randomUUID } from "node:crypto";
 import { type NextRequest } from "next/server";
 
-import { audit } from "@/lib/audit";
 import { ApiError } from "@/lib/api/types";
 import { ok, fail } from "@/lib/api/wrappers";
 import { contactPatchSchema, validateRequest } from "@/lib/schemas";
-import { hashCpf, encryptCpfSql } from "@/lib/contacts/cpf";
 import { createClient } from "@/lib/supabase/server";
-import type { Contact } from "@/lib/types/contacts";
+
+import { getContactHandler, patchContactHandler } from "../_handler";
 
 export const dynamic = "force-dynamic";
-
-const SELECT_COLS =
-  "id, organization_id, name, display_name, email, email_normalized, phone_number, cpf_hash, birthdate, is_blocked, blocked_reason, is_anonymized, anonymized_at, is_merged_into, merged_at, consent, tags, source, source_metadata, created_at, updated_at, last_activity_at";
-
-const ROLE_RANK: Record<string, number> = {
-  viewer: 1,
-  agent: 2,
-  manager: 3,
-  admin: 4,
-};
 
 export async function GET(
   req: NextRequest,
@@ -46,67 +32,25 @@ export async function GET(
     return fail("unauthenticated", "Auth required.", 401, { requestId });
   }
 
-  const { data, error } = await supabase
-    .from("contacts")
-    .select(SELECT_COLS)
-    .eq("id", id)
-    .maybeSingle();
-
-  if (error) {
-    return fail("internal_error", error.message, 500, { requestId });
-  }
-  if (!data) {
-    return fail("not_found", "Contato não encontrado.", 404, { requestId });
-  }
-  const contact = data as Contact;
-
   const decryptPurpose = req.headers.get("x-decrypt-purpose");
-  let cpfDecrypted: string | null = null;
-  let cpfDecryptDenied = false;
 
-  if (decryptPurpose && contact.cpf_hash) {
-    const { data: membership } = await supabase
-      .from("user_organizations")
-      .select("role")
-      .eq("user_id", user.id)
-      .eq("organization_id", contact.organization_id)
-      .is("revoked_at", null)
-      .maybeSingle();
-
-    const role = membership?.role as string | undefined;
-    const rank = role ? (ROLE_RANK[role] ?? 0) : 0;
-    if (rank < ROLE_RANK.manager!) {
-      cpfDecryptDenied = true;
-    } else {
-      const { data: dec, error: decErr } = await supabase.rpc("decrypt_cpf", {
-        p_contact_id: id,
-      });
-      if (decErr) {
-        console.warn("[contacts.get] decrypt_cpf RPC unavailable", decErr.message);
-      } else if (typeof dec === "string") {
-        cpfDecrypted = dec;
-      }
-      await audit({
-        action: "contact.updated", // re-purpose for read-with-decrypt? use dedicated when available
-        actorUserId: user.id,
-        organizationId: contact.organization_id,
-        resourceType: "contact",
-        resourceId: contact.id,
+  try {
+    const result = await getContactHandler(
+      supabase,
+      {
+        organization_id: "",
+        actor: { type: "user", id: user.id },
         requestId,
-        metadata: { decrypt_purpose: decryptPurpose, success: !!cpfDecrypted },
-      });
+      },
+      { contactId: id, decryptPurpose },
+    );
+    return ok(result, { requestId });
+  } catch (err) {
+    if (err instanceof ApiError) {
+      return fail(err.code, err.message, err.status, { requestId });
     }
+    throw err;
   }
-
-  return ok(
-    {
-      ...contact,
-      cpf_available: !!contact.cpf_hash,
-      cpf_decrypted: cpfDecrypted,
-      cpf_decrypt_denied: cpfDecryptDenied || undefined,
-    },
-    { requestId },
-  );
 }
 
 export async function PATCH(
@@ -125,27 +69,6 @@ export async function PATCH(
     return fail("unauthenticated", "Auth required.", 401, { requestId });
   }
 
-  const { data: existing, error: selErr } = await supabase
-    .from("contacts")
-    .select("id, organization_id, is_anonymized")
-    .eq("id", id)
-    .maybeSingle();
-
-  if (selErr) {
-    return fail("internal_error", selErr.message, 500, { requestId });
-  }
-  if (!existing) {
-    return fail("not_found", "Contato não encontrado.", 404, { requestId });
-  }
-  if (existing.is_anonymized) {
-    return fail(
-      "lgpd_anonymization_irreversible",
-      "Contato anonimizado — edição bloqueada (LGPD).",
-      403,
-      { requestId },
-    );
-  }
-
   let input;
   try {
     input = await validateRequest(contactPatchSchema, req);
@@ -159,69 +82,22 @@ export async function PATCH(
     throw err;
   }
 
-  const patch: Record<string, unknown> = {};
-  if (input.name !== undefined) patch.name = input.name;
-  if (input.display_name !== undefined) patch.display_name = input.display_name;
-  if (input.email !== undefined) {
-    patch.email = input.email;
-    patch.email_normalized = input.email ? input.email.trim().toLowerCase() : null;
+  try {
+    const contact = await patchContactHandler(
+      supabase,
+      {
+        organization_id: "",
+        actor: { type: "user", id: user.id },
+        requestId,
+      },
+      id,
+      input,
+    );
+    return ok(contact, { requestId });
+  } catch (err) {
+    if (err instanceof ApiError) {
+      return fail(err.code, err.message, err.status, { requestId });
+    }
+    throw err;
   }
-  if (input.phone_number !== undefined) patch.phone_number = input.phone_number;
-  if (input.birthdate !== undefined) patch.birthdate = input.birthdate;
-  if (input.tags !== undefined) patch.tags = input.tags;
-  if (input.source !== undefined) patch.source = input.source;
-  if (input.source_metadata !== undefined) patch.source_metadata = input.source_metadata;
-  if (input.consent !== undefined) patch.consent = input.consent;
-  if (input.cpf !== undefined) {
-    patch.cpf_hash = hashCpf(input.cpf);
-    const enc = await encryptCpfSql(supabase, input.cpf);
-    if (enc) patch.cpf_encrypted = enc;
-  }
-
-  if (Object.keys(patch).length === 0) {
-    return fail("invalid_request", "Nenhum campo para atualizar.", 400, { requestId });
-  }
-
-  patch.updated_at = new Date().toISOString();
-
-  const { data: updated, error: updErr } = await supabase
-    .from("contacts")
-    .update(patch)
-    .eq("id", id)
-    .select(SELECT_COLS)
-    .maybeSingle();
-
-  if (updErr) {
-    return fail("internal_error", updErr.message, 500, { requestId });
-  }
-  if (!updated) {
-    return fail("not_found", "Contato não encontrado após update.", 404, { requestId });
-  }
-
-  const contact = updated as Contact;
-
-  await supabase
-    .rpc("emit_event", {
-      p_event_type: "contact.updated",
-      p_entity_kind: "contact",
-      p_entity_id: contact.id,
-      p_payload: { fields: Object.keys(patch).filter((k) => k !== "updated_at") },
-      p_metadata: { request_id: requestId, actor_user_id: user.id },
-      p_organization_id: contact.organization_id,
-    })
-    .then(({ error }) => {
-      if (error) console.error("[contacts.patch] emit_event failed", error.message);
-    });
-
-  await audit({
-    action: "contact.updated",
-    actorUserId: user.id,
-    organizationId: contact.organization_id,
-    resourceType: "contact",
-    resourceId: contact.id,
-    requestId,
-    metadata: { fields: Object.keys(patch).filter((k) => k !== "updated_at") },
-  });
-
-  return ok(contact, { requestId });
 }
