+++ app/api/v1/conversations/_handler.ts
@@ -1,0 +1,220 @@
+/**
+ * Core handlers para /api/v1/conversations.
+ *
+ * Reusados pelo Route Handler REST e por MCP tools (S-13.03/04).
+ */
+import type { SupabaseClient } from "@supabase/supabase-js";
+
+import { ApiError } from "@/lib/api/types";
+import type { Actor, HandlerCtx } from "@/lib/api/handlers/types";
+import { audit } from "@/lib/audit";
+import type {
+  ListConversationsQuery,
+  UpdateConversationStatusInput,
+} from "@/lib/schemas";
+import type { Conversation } from "@/lib/types/messaging";
+
+type SB = SupabaseClient;
+
+const SELECT_COLS = `
+  id, organization_id, contact_id, channel_session_id, channel, status,
+  status_changed_at, assigned_to_user_id, assigned_at, last_inbound_at,
+  last_outbound_at, last_message_at, last_message_preview,
+  unread_count_for_assignee, is_group, group_chat_id, metadata,
+  created_at, updated_at,
+  contacts:contact_id (id, display_name, name, phone_number, is_anonymized, tags, is_blocked)
+`;
+
+interface CursorPayload {
+  last_message_at: string | null;
+  id: string;
+}
+
+function encodeCursor(p: CursorPayload): string {
+  return Buffer.from(JSON.stringify(p), "utf8").toString("base64url");
+}
+function decodeCursor(raw: string): CursorPayload | null {
+  try {
+    const json = Buffer.from(raw, "base64url").toString("utf8");
+    const parsed = JSON.parse(json) as CursorPayload;
+    if (typeof parsed.id !== "string") return null;
+    return parsed;
+  } catch {
+    return null;
+  }
+}
+
+function actorAuditPayload(actor: Actor): {
+  actorUserId: string | null;
+  metadataActor: Record<string, unknown>;
+} {
+  if (actor.type === "user") {
+    return { actorUserId: actor.id, metadataActor: { actor_type: "user" } };
+  }
+  return {
+    actorUserId: null,
+    metadataActor: {
+      actor_type: "ai_agent",
+      actor_id: actor.id,
+      ...(actor.api_token_id ? { actor_api_token_id: actor.api_token_id } : {}),
+    },
+  };
+}
+
+// ---------------------------------------------------------------------------
+// list
+// ---------------------------------------------------------------------------
+
+export interface ListConversationsResult {
+  conversations: Conversation[];
+  cursor: string | null;
+  has_more: boolean;
+}
+
+export async function listConversationsHandler(
+  supabase: SB,
+  ctx: HandlerCtx,
+  q: ListConversationsQuery,
+): Promise<ListConversationsResult> {
+  let query = supabase
+    .from("conversations")
+    .select(SELECT_COLS)
+    .order("last_message_at", { ascending: false, nullsFirst: false })
+    .order("id", { ascending: false })
+    .limit(q.limit + 1);
+
+  if (q.status) query = query.eq("status", q.status);
+  if (q.channel_session_id) query = query.eq("channel_session_id", q.channel_session_id);
+
+  if (q.assigned_to === "me") {
+    if (ctx.actor.type !== "user") {
+      throw new ApiError(
+        400,
+        "invalid_request",
+        undefined,
+        ctx.requestId,
+        '"assigned_to=me" requer ator humano.',
+      );
+    }
+    query = query.eq("assigned_to_user_id", ctx.actor.id);
+  } else if (q.assigned_to === "unassigned") {
+    query = query.is("assigned_to_user_id", null);
+  } else if (q.assigned_to) {
+    query = query.eq("assigned_to_user_id", q.assigned_to);
+  }
+
+  if (q.search) {
+    const s = q.search.trim().replace(/[%_]/g, (m) => `\\${m}`);
+    query = query.ilike("last_message_preview", `%${s}%`);
+  }
+
+  if (q.cursor) {
+    const c = decodeCursor(q.cursor);
+    if (!c) {
+      throw new ApiError(400, "invalid_cursor", undefined, ctx.requestId, "Cursor inválido.");
+    }
+    if (c.last_message_at) {
+      query = query.or(
+        `last_message_at.lt.${c.last_message_at},and(last_message_at.eq.${c.last_message_at},id.lt.${c.id})`,
+      );
+    } else {
+      query = query.is("last_message_at", null).lt("id", c.id);
+    }
+  }
+
+  const { data, error } = await query;
+  if (error) {
+    throw new ApiError(500, "internal_error", undefined, ctx.requestId, error.message);
+  }
+
+  const rows = (data ?? []) as unknown as Conversation[];
+  const hasMore = rows.length > q.limit;
+  const page = hasMore ? rows.slice(0, q.limit) : rows;
+  const last = page[page.length - 1];
+  const cursor =
+    hasMore && last
+      ? encodeCursor({ last_message_at: last.last_message_at, id: last.id })
+      : null;
+
+  return { conversations: page, cursor, has_more: hasMore };
+}
+
+// ---------------------------------------------------------------------------
+// get
+// ---------------------------------------------------------------------------
+
+export async function getConversationHandler(
+  supabase: SB,
+  ctx: HandlerCtx,
+  conversationId: string,
+): Promise<Conversation> {
+  const { data, error } = await supabase
+    .from("conversations")
+    .select(SELECT_COLS)
+    .eq("id", conversationId)
+    .maybeSingle();
+
+  if (error) {
+    throw new ApiError(500, "internal_error", undefined, ctx.requestId, error.message);
+  }
+  if (!data) {
+    throw new ApiError(404, "not_found", undefined, ctx.requestId, "Conversa não encontrada.");
+  }
+  return data as unknown as Conversation;
+}
+
+// ---------------------------------------------------------------------------
+// update status (claim/close/release)
+// ---------------------------------------------------------------------------
+
+export async function updateConversationStatusHandler(
+  supabase: SB,
+  ctx: HandlerCtx,
+  conversationId: string,
+  input: UpdateConversationStatusInput,
+): Promise<Conversation> {
+  const update: Record<string, unknown> = {
+    status: input.status,
+    status_changed_at: new Date().toISOString(),
+  };
+  // Atalho: status='claimed' assume o atendimento se ator for usuário humano.
+  if (input.status === "claimed" && ctx.actor.type === "user") {
+    update.assigned_to_user_id = ctx.actor.id;
+    update.assigned_at = new Date().toISOString();
+  }
+
+  const { data, error } = await supabase
+    .from("conversations")
+    .update(update)
+    .eq("id", conversationId)
+    .select(SELECT_COLS)
+    .maybeSingle();
+
+  if (error) {
+    throw new ApiError(500, "internal_error", undefined, ctx.requestId, error.message);
+  }
+  if (!data) {
+    throw new ApiError(404, "not_found", undefined, ctx.requestId, "Conversa não encontrada.");
+  }
+
+  const conv = data as unknown as Conversation;
+  const action =
+    input.status === "claimed"
+      ? "conversation.claimed"
+      : input.status === "closed"
+        ? "conversation.closed"
+        : "conversation.released";
+
+  const a = actorAuditPayload(ctx.actor);
+  await audit({
+    action,
+    actorUserId: a.actorUserId,
+    organizationId: conv.organization_id,
+    resourceType: "conversation",
+    resourceId: conv.id,
+    requestId: ctx.requestId,
+    metadata: { ...a.metadataActor, status: input.status },
+  });
+
+  return conv;
+}
